Changes to allow compiling in Windows

Former-commit-id: 30f570be8a776614e868029229a965198da6e16d

+++ tests/test-cnn.cc
@@ -24,7 +24,7 @@ BOOST_GLOBAL_FIXTURE(ConfigureCNNTest);
 BOOST_AUTO_TEST_CASE( aligned_allocator ) {
   cnn::CPUAllocator a;
   void* mem = a.malloc(1024);
-  BOOST_CHECK_EQUAL(((unsigned long)(mem) & 0x1f), 0);
+  BOOST_CHECK_EQUAL(((uintptr_t)(mem) & 0x1f), 0);
   ((char*)mem)[0] = 99;
   a.zero(mem, 1024);
   BOOST_CHECK_EQUAL(((char*)mem)[0], 0);
